Rebalance classes
